Uncommenting rover_common_launch, ready for testing
